+++ src/Viewer/App/Spec.ts
@@ -19,7 +19,8 @@ namespace LiteMol.Viewer {
             'molecule.coordinateStreaming.defaultRadius': 10,
             'density.defaultVisualBehaviourRadius': 5,
 
-            'extensions.densityStreaming.defaultServer': 'https://webchem.ncbr.muni.cz/DensityServer/' // 'http://localhost:1337/DensityServer/'
+            'extensions.densityStreaming.defaultServer': 'https://webchem.ncbr.muni.cz/DensityServer/', // 'http://localhost:1337/DensityServer/'
+            'extensions.rnaLoops.defaultServer': 'http://rna.bgsu.edu/rna3dhub/loops/download/#id'
         },
         transforms: [
             // Root transforms -- things that load data.
@@ -80,6 +81,9 @@ namespace LiteMol.Viewer {
             // annotations
             { transformer: PDBe.SequenceAnnotation.DownloadAndCreate, view: Views.Transform.Empty, initiallyCollapsed: true },
             { transformer: PDBe.SequenceAnnotation.CreateSingle, view: PDBe.Views.CreateSequenceAnnotationView, initiallyCollapsed: true },
+
+            { transformer: Extensions.RNALoops.DownloadAndCreate, view: Extensions.RNALoops.CreateLoopAnnotationView, initiallyCollapsed: true },
+            { transformer: Extensions.RNALoops.ApplyTheme, view: Views.Transform.Empty, initiallyCollapsed: false },
         ],
         behaviours: [
             // you will find the source of all behaviours in the Bootstrap/Behaviour directory