+++ src/lib/Visualization/Molecule/Cartoons/GeometryElements.ts
@@ -9,7 +9,9 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
         turnWidth = 0.1;
 
         strandWidth = 0.15;
-        strandLineWidth = 0.1;
+        
+        nucleotideStrandLineWidth = 0.15;
+        nucleotideStrandFactor = 3;
 
         helixWidth = 1.1;
         helixHeight = 0.1;
@@ -42,6 +44,15 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
 
         gapsBuilder = GB.createDynamic(256, 512);
 
+        private dCones = GB.createDynamic(1, 1);
+        private dConesInit = false;
+        get directionConesBuilder() {
+            if (this.dConesInit) return this.dCones;
+            this.dConesInit = true;
+            this.dCones = GB.createDynamic(this.residueCount, this.residueCount);
+            return this.dCones;
+        }
+
         translationMatrix: THREE.Matrix4 = new THREE.Matrix4();
         scaleMatrix: THREE.Matrix4 = new THREE.Matrix4();
         rotationMatrix: THREE.Matrix4 = new THREE.Matrix4();
@@ -69,7 +80,7 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
     }
 
     function makeStrandLineTemplate(ctx: Context) {
-        let radius = ctx.params.strandLineWidth, tessalation = ctx.params.tessalation;
+        let radius = ctx.params.nucleotideStrandLineWidth, tessalation = ctx.params.tessalation;
         let capPoints = 0, radiusPoints = 0, geom: THREE.Geometry;
 
         switch (tessalation) {
@@ -96,20 +107,12 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
             index: (<any>templ.attributes).index.array,
             geometry: templ
         };
-
-        let atoms = ctx.model.data.atoms, residues = ctx.model.data.residues;
-        let positions = ctx.model.positions;
-        ctx.strandArrays = {
-            startIndex: residues.atomStartIndex,
-            endIndex: residues.atomEndIndex,
-            x: positions.x, y: positions.y, z: positions.z,
-            name: atoms.name
-        };
     }
 
     export function buildUnit(unit: CartoonAsymUnit, ctx: Context) {
         let state = ctx.state, params = ctx.params;
         let builder = ctx.builder;
+
         for (let index = 0, _max = unit.residueCount; index < _max; index++) {
             state.vertexMap.startElement(unit.residueIndex[index]);
             let numVertices = state.verticesDone;
@@ -119,7 +122,8 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
             if (ctx.isTrace || unit.backboneOnly) {
                 switch (unit.residueType[index]) {
                     case Core.Structure.SecondaryStructureType.Strand:
-                        builder.addTube(unit, state, params.strandWidth, params.strandWidth);
+                        builder.addTube(unit, state, params.strandWidth, params.strandWidth, 
+                            builder.hasP(unit.residueIndex[index], ctx.strandArrays) ? params.nucleotideStrandFactor : 1);
                         if (start || end) {
                             builder.addTubeCap(unit, state, params.strandWidth, params.strandWidth, start, end);
                         }
@@ -130,7 +134,7 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
                         builder.addStrandLine(unit, state, ctx.strandTemplate, ctx.strandArrays, unit.residueIndex[index]);
                         break;
                     default:
-                        builder.addTube(unit, state, params.turnWidth, params.turnWidth);
+                        builder.addTube(unit, state, params.turnWidth, params.turnWidth, params.turnWidth);
                         if (start || end) {
                             builder.addTubeCap(unit, state, params.turnWidth, params.turnWidth, start, end);
                         }
@@ -139,7 +143,7 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
             } else {
                 switch (unit.residueType[index]) {
                     case Core.Structure.SecondaryStructureType.Helix:
-                        builder.addTube(unit, state, params.helixWidth, params.helixHeight);
+                        builder.addTube(unit, state, params.helixWidth, params.helixHeight, 1);
                         if (start) {
                             builder.addTubeCap(unit, state, params.helixWidth, params.helixHeight, true, false);
                         } else if (end) {
@@ -154,7 +158,8 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
                         }
                         break;
                     case Core.Structure.SecondaryStructureType.Strand:
-                        builder.addTube(unit, state, params.strandWidth, params.strandWidth);
+                        builder.addTube(unit, state, params.strandWidth, params.strandWidth, 
+                            builder.hasP(unit.residueIndex[index], ctx.strandArrays) ? params.nucleotideStrandFactor : 1);
                         if (start || end) {
                             builder.addTubeCap(unit, state, params.strandWidth, params.strandWidth, start, end);
                         }
@@ -165,12 +170,15 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
                         builder.addStrandLine(unit, state, ctx.strandTemplate, ctx.strandArrays, unit.residueIndex[index]);
                         break;
                     default:
-                        builder.addTube(unit, state, params.turnWidth, params.turnWidth);
+                        builder.addTube(unit, state, params.turnWidth, params.turnWidth, 1);
                         if (start || end) {
                             builder.addTubeCap(unit, state, params.turnWidth, params.turnWidth, start, end);
                         }
                         break;
                 }
+            }
+            if (ctx.parameters.showDirectionCones && unit.residueType[index] !== Core.Structure.SecondaryStructureType.Strand) {
+                renderDirectionCone(ctx, unit, 2 * params.sheetHeight, index);
             }
 
             state.vertexMap.addVertexRange(numVertices, state.verticesDone);
@@ -184,12 +192,45 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
     }
 
     import Vec3 = Core.Geometry.LinearAlgebra.Vector3
+    import Mat4 = Core.Geometry.LinearAlgebra.Matrix4
     function renderGap(ctx: Context, unitA: CartoonAsymUnit, unitB: CartoonAsymUnit) {
         const aL = unitA.controlPoints.length;
         const cpA = unitA.controlPoints, cpB = unitB.controlPoints;
         const a = Vec3.fromValues(cpA[aL - 3], cpA[aL - 2], cpA[aL - 1]), b = Vec3.fromValues(cpB[0], cpB[1], cpB[2]);
         const r = ctx.state.params.turnWidth / 2;
         GB.addDashedLine(ctx.state.gapsBuilder, a, b, 0.5, 0.5, r);
+    }
+
+    const coneTemplate = (function() {
+        const geom = new THREE.CylinderGeometry(0, 1, 1, 6, 1);
+        const ret = GeometryHelper.toRawGeometry(geom);
+        geom.dispose();
+        return ret;
+    })();
+    const coneDirection = Vec3.zero(), coneUp = Vec3.fromValues(0, 1, 0), coneA = Vec3.zero(), coneB = Vec3.zero(), coneTranslation = Vec3.zero(), coneScale = Vec3.zero(), coneRotation = Mat4.identity();
+
+    function renderDirectionCone(ctx: Context, unit: CartoonAsymUnit, radius: number, residueIndex: number) {
+        if (unit.residueCount <= 2) return;
+
+        const cp = unit.controlPoints;
+        const i = residueIndex * unit.linearSegmentCount + ((0.35 * unit.linearSegmentCount + 1) | 0);
+        const j = residueIndex * unit.linearSegmentCount + ((0.85 * unit.linearSegmentCount + 1) | 0);
+
+        if (i === j || 3 * j > cp.length) return;
+
+        Vec3.set(coneTranslation, cp[3 * i], cp[3 * i + 1], cp[3 * i + 2]);        
+        Vec3.set(coneA, cp[3 * i], cp[3 * i + 1], cp[3 * i + 2]);
+        Vec3.set(coneB, cp[3 * j], cp[3 * j + 1], cp[3 * j + 2]);
+        Vec3.sub(coneA, coneB, coneA);
+
+        const l = Vec3.magnitude(coneA);
+        if (l <= 0.1) return;
+
+        Vec3.set(coneScale, 2 * radius, l, 2 * radius);
+        Vec3.normalize(coneA, coneA);
+        Vec3.makeRotation(coneRotation, coneUp, coneA);
+
+        GB.addRawTransformed(ctx.state.directionConesBuilder, coneTemplate, coneScale, coneTranslation, coneRotation);
     }
 
     export async function buildUnitsAsync(ctx: Context): Promise<void> {
@@ -235,6 +276,10 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
 
         if (state.gapsBuilder.vertices.elementCount) {
             ctx.geom.gapsGeometry = GB.toBufferGeometry(state.gapsBuilder);
+        }
+
+        if (state.directionConesBuilder.vertices.elementCount) {
+            ctx.geom.directionConesGeometry = GB.toBufferGeometry(state.directionConesBuilder);
         }
 
         let map = ctx.geom.vertexMap,
@@ -286,7 +331,7 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
             return v;
         }
 
-        addTube(element: CartoonAsymUnit, state: CartoonsGeometryState, width: number, height: number) {
+        addTube(element: CartoonAsymUnit, state: CartoonsGeometryState, width: number, height: number, waveFactor: number) {
             let verticesDone = state.verticesDone,
                 i = 0, j = 0,
                 radialVector = this.tempVectors[0], normalVector = this.tempVectors[1],
@@ -300,21 +345,26 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
                 normalVectors = element.normalVectors,
                 radialSegmentCount = state.params.radialSegmentCount;
 
+            const di = 1 / (elementOffsetEnd - elementOffsetStart);
+
             for (i = elementOffsetStart; i <= elementOffsetEnd; i++) {
                 this.setVector(torsionVectors, i, u);
                 this.setVector(normalVectors, i, v);
 
-                for (j = 0; j < radialSegmentCount; j++) {
+                const tt = di * (i - elementOffsetStart) - 0.5;
+                const ff = 1 + (waveFactor - 1) * (Math.cos(2 * Math.PI * tt) + 1);
+                const w = ff * width, h = ff * height;
 
+                for (j = 0; j < radialSegmentCount; j++) {
                     let t = 2 * Math.PI * j / radialSegmentCount;
 
                     a.copy(u);
                     b.copy(v);
-                    radialVector.addVectors(a.multiplyScalar(width * Math.cos(t)), b.multiplyScalar(height * Math.sin(t)));
+                    radialVector.addVectors(a.multiplyScalar(w * Math.cos(t)), b.multiplyScalar(h * Math.sin(t)));
 
                     a.copy(u);
                     b.copy(v);
-                    normalVector.addVectors(a.multiplyScalar(height * Math.cos(t)), b.multiplyScalar(width * Math.sin(t)));
+                    normalVector.addVectors(a.multiplyScalar(h * Math.cos(t)), b.multiplyScalar(w * Math.sin(t)));
                     normalVector.normalize();
 
                     this.setVector(elementPoints, i, tempPos);
@@ -515,7 +565,6 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
         }
 
         private findN3(index: number, arrays: { startIndex: number[]; endIndex: number[]; x: number[]; y: number[]; z: number[]; name: string[] }, target: THREE.Vector3) {
-
             let start = arrays.startIndex[index], end = arrays.endIndex[index];
             let found = false;
 
@@ -528,6 +577,14 @@ namespace LiteMol.Visualization.Molecule.Cartoons.Geometry {
             }
 
             return found;
+        }
+
+        hasP(index: number, arrays: { startIndex: number[]; endIndex: number[]; x: number[]; y: number[]; z: number[]; name: string[] }) {
+            let start = arrays.startIndex[index], end = arrays.endIndex[index];
+            for (let i = start; i < end; i++) {
+                if (arrays.name[i] === "P") return true;
+            }
+            return false;
         }
 
         addStrandLine(