+++ src/lib/Visualization/CHANGELOG.md
@@ -1,4 +1,11 @@
-﻿# 1.7.5
+﻿# 1.7.7
+* Added optional "direction cones" to cartoons.
+
+# 1.7.6
+* Swapped direction of the camera Z axis.
+* DNA/RNA now render as "waves".
+
+# 1.7.5
 * Fixed secondary structure assigment for cartoon models.
 
 # 1.7.4